requirements.txt
Django==4.2.7
psycopg2-binary
PyJWT==2.8.0
pillow==10.4.0
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
